requirements.txt
Django==3.0.4
Markdown==3.1.1
Whoosh==2.7.4
jieba==0.42.1
django_ckeditor==5.9.0
django_haystack==2.8.1
